fix a bug when attack_radius < 0

+++ Classes/core/entity/entity.h
@@ -59,7 +59,7 @@ struct Buff {
     // actual_defence = base_defence + defence_correction_
     Defence defence_correction_{0, 0};
     // actual_attack_radius = base_attack_radius + attack_radius
-    size_t attack_radius_ = 0;
+    int32_t attack_radius_ = 0;
     bool invincible_ = false;
     bool silent_ = false;
     // actual_inspiration_strike = base_attack * (1 + attack) *
@@ -71,7 +71,7 @@ struct Buff {
     BUFF_CONSTUCTOR(double, attack)
     BUFF_CONSTUCTOR(double, real_attack)
     BUFF_CONSTUCTOR(Defence, defence_correction)
-    BUFF_CONSTUCTOR(size_t, attack_radius)
+    BUFF_CONSTUCTOR(int32_t, attack_radius)
     BUFF_CONSTUCTOR(bool, invincible)
     BUFF_CONSTUCTOR(bool, silent)
     BUFF_CONSTUCTOR(double, inspiration_strike)
@@ -85,7 +85,7 @@ struct Buff {
     constexpr Buff() = default;
     constexpr Buff(int32_t attack_speed, double speed, double attack,
                    double real_attack, Defence defence_correction,
-                   size_t attack_radius, bool invincible,
+                   int32_t attack_radius, bool invincible,
                    bool silent, double inspiration_strike)
         : attack_speed_(attack_speed), speed_(speed), attack_(attack),
           real_attack_(real_attack), defence_correction_(defence_correction),
